Add new copy directory method and light refactoring
Add new copy directory method and light refactoring
