+++ BN - Primitive Launcher/BN - Primitive Launcher.csproj
@@ -88,7 +88,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FolderPicker.cs" />
+    <Compile Include="Classes\FolderPicker.cs" />
+    <Compile Include="Classes\Utils.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
